LayoutTests/platform/ios-simulator/editing/pasteboard/file-input-files-access-expected.txt

CONSOLE MESSAGE: line 28: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(files)')

Tests for multi-file drag onto file input elements for https://bugs.webkit.org/show_bug.cgi?id=25862

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Dragging a single (non-existant) file to a file input control:
PASS successfullyParsed is true

TEST COMPLETE
